modal for user profile finished & solved bugs
slight changes to folder structure and common components
