+++ components/App/ArbSwap/Dashboard/Offers/OffersSent/index.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function OffersSent() {
+  return <div className="text-black">Offers Sent</div>
+}
